requirements.txt
yt-dlp
ffmpeg-python
openai-whisper
streamlit
mediapipe
